requirements.txt
streamlit==1.37.1
tensorflow==2.19.0
numpy==1.26.4
pillow==10.4.0